osx: 64bit: fix `cnt` and `size` variables warnings

+++ Source/OS9exec_core/consio.h
@@ -95,7 +95,7 @@ extern short TermWinScrolls;
 extern ttydev_typ main_mco;
 
 extern Boolean KeyToBuffer(ttydev_typ *mco, char key);
-extern Boolean DevReady(long *count);
+extern Boolean DevReady(uint32_t *count);
 #endif
 
 /* eof */
